In some cases we want to have counter values with more precise values than just integers.
For instance, an average duration for a very fast operation (like merging locations) might be taking less than a millisecond. This change allows tracing the counters as doubles as well as just longs.

Before: 'MergeContentLocation.AverageMs: 0"
After: 'MergeContentLocation.AverageMs: 0.005"

+++ Public/Src/Cache/ContentStore/Hashing/ContentHasher.cs
@@ -268,7 +268,7 @@ namespace BuildXL.Cache.ContentStore.Hashing
             if (_calls > 0)
             {
                 counters.Add("ContentHasherCalls", Interlocked.Read(ref _calls));
-                counters.Add("ContentHasherMs", (long)new TimeSpan(Interlocked.Read(ref _ticks)).TotalMilliseconds);
+                counters.Add("ContentHasherMs", new TimeSpan(Interlocked.Read(ref _ticks)).TotalMilliseconds);
                 counters.Add("ContentHasherMaxConcurrency", _algorithmsPool.Size);
             }
